requirements.txt
Django==3.1.7
django-ckeditor==6.0.0
django-resized==0.3.11
Pillow==8.2.0